修改最新的FSS

+++ 03_Abstract（组件接口）/FOPS.Abstract/Fss/Entity/TaskGroupVO.cs
@@ -65,7 +65,7 @@ namespace FOPS.Abstract.Fss.Entity
         /// <summary>
         /// 运行平均耗时
         /// </summary>
-        public int RunSpeedAvg { get; set; }
+        public long RunSpeedAvg { get; set; }
         
         /// <summary>
         /// 运行次数
